chore: move SubgraphDeploymentID tests to types.rs
